An example of XY maps.
This NB computes XY maps using run 6364 to correct energy and
lifetime.

It shows that time dependence does not seem to affect the attainable
resolution in Krypton in spite of being substantial.

It also shows that geometrical maps can be computed on the flight. No
gain is attained to use high resolution maps.